move register_tools function into rdkit_mcp
