+++ rdkit_mcp/register_tools.py
