+++ run_server.py
@@ -3,7 +3,7 @@ import asyncio
 import logging
 from mcp.server.fastmcp import FastMCP
 
-from register_tools import register_tools
+from rdkit_mcp.register_tools import register_tools
 import yaml
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
